Add GHC primitives
Most primitives from ghc-prim are relatively simple to implement:
as we can already convert values from/to an AST, all that is left
is to simply call the primitive function on the converted values.

For other prims this is more involved, e.g. ByteArray prims which
live in IO, or primitives on W/W created bindings. These have to
be implemented specially to ensure the type of the result is correct.